Youtube_Video_Summarizer/requirements.txt
gradio
yt-dlp
openai-whisper
torch
langchain-community
sentence-transformers
ollama
chromadb